Steglib-main 2/requirements.txt
albumentations
catalyst
efficientnet_pytorch
numpy
pandas
scipy
scikit-image
torch
tqdm
tensorboard
matplotlib
ipywidgets
flask
flask-session
python-dotenv
tensorflow
efficientnet
jpegio
seaborn==0.11.2